+++ src/minecraft/net/minecraft/block/BlockDropper.java
@@ -50,7 +50,7 @@ public class BlockDropper extends BlockDispenser
                 {
                     EnumFacing enumfacing = worldIn.getBlockState(pos).getValue(FACING);
                     BlockPos blockpos = pos.offset(enumfacing);
-                    IInventory iinventory = TileEntityHopper.getInventoryAtPosition(worldIn, (double)blockpos.getX(), (double)blockpos.getY(), (double)blockpos.getZ());
+                    IInventory iinventory = TileEntityHopper.getInventoryAtPosition(worldIn, blockpos.getX(), blockpos.getY(), blockpos.getZ());
                     ItemStack itemstack1;
 
                     if (iinventory == null)
