+++ src/minecraft/net/minecraft/block/BlockLeaves.java
@@ -193,9 +193,9 @@ public abstract class BlockLeaves extends BlockLeavesBase
     {
         if (worldIn.isRainingAt(pos.up()) && !World.doesBlockHaveSolidTopSurface(worldIn, pos.down()) && rand.nextInt(15) == 1)
         {
-            double d0 = (double)((float)pos.getX() + rand.nextFloat());
+            double d0 = (float)pos.getX() + rand.nextFloat();
             double d1 = (double)pos.getY() - 0.05D;
-            double d2 = (double)((float)pos.getZ() + rand.nextFloat());
+            double d2 = (float)pos.getZ() + rand.nextFloat();
             worldIn.spawnParticle(EnumParticleTypes.DRIP_WATER, d0, d1, d2, 0.0D, 0.0D, 0.0D);
         }
     }
