+++ src/minecraft/net/minecraft/block/BlockPortal.java
@@ -42,7 +42,6 @@ public class BlockPortal extends BlockBreakable
 
             for (blockpos = pos; !World.doesBlockHaveSolidTopSurface(worldIn, blockpos) && blockpos.getY() > 0; blockpos = blockpos.down())
             {
-                ;
             }
 
             if (i > 0 && !worldIn.getBlockState(blockpos.up()).getBlock().isNormalCube())
@@ -233,9 +232,9 @@ public class BlockPortal extends BlockBreakable
 
         for (int i = 0; i < 4; ++i)
         {
-            double d0 = (double)((float)pos.getX() + rand.nextFloat());
-            double d1 = (double)((float)pos.getY() + rand.nextFloat());
-            double d2 = (double)((float)pos.getZ() + rand.nextFloat());
+            double d0 = (float)pos.getX() + rand.nextFloat();
+            double d1 = (float)pos.getY() + rand.nextFloat();
+            double d2 = (float)pos.getZ() + rand.nextFloat();
             double d3 = ((double)rand.nextFloat() - 0.5D) * 0.5D;
             double d4 = ((double)rand.nextFloat() - 0.5D) * 0.5D;
             double d5 = ((double)rand.nextFloat() - 0.5D) * 0.5D;
@@ -244,12 +243,12 @@ public class BlockPortal extends BlockBreakable
             if (worldIn.getBlockState(pos.west()).getBlock() != this && worldIn.getBlockState(pos.east()).getBlock() != this)
             {
                 d0 = (double)pos.getX() + 0.5D + 0.25D * (double)j;
-                d3 = (double)(rand.nextFloat() * 2.0F * (float)j);
+                d3 = rand.nextFloat() * 2.0F * (float)j;
             }
             else
             {
                 d2 = (double)pos.getZ() + 0.5D + 0.25D * (double)j;
-                d5 = (double)(rand.nextFloat() * 2.0F * (float)j);
+                d5 = rand.nextFloat() * 2.0F * (float)j;
             }
 
             worldIn.spawnParticle(EnumParticleTypes.PORTAL, d0, d1, d2, d3, d4, d5);
@@ -365,7 +364,6 @@ public class BlockPortal extends BlockBreakable
 
             for (BlockPos blockpos = p_i45694_2_; p_i45694_2_.getY() > blockpos.getY() - 21 && p_i45694_2_.getY() > 0 && this.func_150857_a(worldIn.getBlockState(p_i45694_2_.down()).getBlock()); p_i45694_2_ = p_i45694_2_.down())
             {
-                ;
             }
 
             int i = this.func_180120_a(p_i45694_2_, this.field_150863_d) - 1;
