+++ src/minecraft/net/minecraft/block/BlockRedstoneRepeater.java
@@ -117,8 +117,8 @@ public class BlockRedstoneRepeater extends BlockRedstoneDiode
             }
 
             f = f / 16.0F;
-            double d3 = (double)(f * (float)enumfacing.getFrontOffsetX());
-            double d4 = (double)(f * (float)enumfacing.getFrontOffsetZ());
+            double d3 = f * (float)enumfacing.getFrontOffsetX();
+            double d4 = f * (float)enumfacing.getFrontOffsetZ();
             worldIn.spawnParticle(EnumParticleTypes.REDSTONE, d0 + d3, d1, d2 + d4, 0.0D, 0.0D, 0.0D);
         }
     }
