+++ src/minecraft/net/minecraft/client/particle/EntityCrit2FX.java
@@ -17,9 +17,9 @@ public class EntityCrit2FX extends EntityFX
     protected EntityCrit2FX(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i46285_8_, double p_i46285_10_, double p_i46285_12_, float p_i46285_14_)
     {
         super(worldIn, xCoordIn, yCoordIn, zCoordIn, 0.0D, 0.0D, 0.0D);
-        this.motionX *= (double)0.1F;
-        this.motionY *= (double)0.1F;
-        this.motionZ *= (double)0.1F;
+        this.motionX *= 0.1F;
+        this.motionY *= 0.1F;
+        this.motionZ *= 0.1F;
         this.motionX += p_i46285_8_ * 0.4D;
         this.motionY += p_i46285_10_ * 0.4D;
         this.motionZ += p_i46285_12_ * 0.4D;
@@ -62,15 +62,15 @@ public class EntityCrit2FX extends EntityFX
         this.moveEntity(this.motionX, this.motionY, this.motionZ);
         this.particleGreen = (float)((double)this.particleGreen * 0.96D);
         this.particleBlue = (float)((double)this.particleBlue * 0.9D);
-        this.motionX *= (double)0.7F;
-        this.motionY *= (double)0.7F;
-        this.motionZ *= (double)0.7F;
-        this.motionY -= (double)0.02F;
+        this.motionX *= 0.7F;
+        this.motionY *= 0.7F;
+        this.motionZ *= 0.7F;
+        this.motionY -= 0.02F;
 
         if (this.onGround)
         {
-            this.motionX *= (double)0.7F;
-            this.motionZ *= (double)0.7F;
+            this.motionX *= 0.7F;
+            this.motionZ *= 0.7F;
         }
     }
