+++ src/minecraft/net/minecraft/client/particle/EntityNoteFX.java
@@ -17,9 +17,9 @@ public class EntityNoteFX extends EntityFX
     protected EntityNoteFX(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn, double p_i1217_8_, double p_i1217_10_, double p_i1217_12_, float p_i1217_14_)
     {
         super(worldIn, xCoordIn, yCoordIn, zCoordIn, 0.0D, 0.0D, 0.0D);
-        this.motionX *= (double)0.01F;
-        this.motionY *= (double)0.01F;
-        this.motionZ *= (double)0.01F;
+        this.motionX *= 0.01F;
+        this.motionY *= 0.01F;
+        this.motionZ *= 0.01F;
         this.motionY += 0.2D;
         this.particleRed = MathHelper.sin(((float)p_i1217_8_ + 0.0F) * (float)Math.PI * 2.0F) * 0.65F + 0.35F;
         this.particleGreen = MathHelper.sin(((float)p_i1217_8_ + 0.33333334F) * (float)Math.PI * 2.0F) * 0.65F + 0.35F;
@@ -65,14 +65,14 @@ public class EntityNoteFX extends EntityFX
             this.motionZ *= 1.1D;
         }
 
-        this.motionX *= (double)0.66F;
-        this.motionY *= (double)0.66F;
-        this.motionZ *= (double)0.66F;
+        this.motionX *= 0.66F;
+        this.motionY *= 0.66F;
+        this.motionZ *= 0.66F;
 
         if (this.onGround)
         {
-            this.motionX *= (double)0.7F;
-            this.motionZ *= (double)0.7F;
+            this.motionX *= 0.7F;
+            this.motionZ *= 0.7F;
         }
     }
