+++ src/minecraft/net/minecraft/client/particle/EntityRainFX.java
@@ -13,9 +13,9 @@ public class EntityRainFX extends EntityFX
     protected EntityRainFX(World worldIn, double xCoordIn, double yCoordIn, double zCoordIn)
     {
         super(worldIn, xCoordIn, yCoordIn, zCoordIn, 0.0D, 0.0D, 0.0D);
-        this.motionX *= (double)0.3F;
+        this.motionX *= 0.3F;
         this.motionY = Math.random() * (double)0.2F + (double)0.1F;
-        this.motionZ *= (double)0.3F;
+        this.motionZ *= 0.3F;
         this.particleRed = 1.0F;
         this.particleGreen = 1.0F;
         this.particleBlue = 1.0F;
@@ -33,11 +33,11 @@ public class EntityRainFX extends EntityFX
         this.prevPosX = this.posX;
         this.prevPosY = this.posY;
         this.prevPosZ = this.posZ;
-        this.motionY -= (double)this.particleGravity;
+        this.motionY -= this.particleGravity;
         this.moveEntity(this.motionX, this.motionY, this.motionZ);
-        this.motionX *= (double)0.98F;
-        this.motionY *= (double)0.98F;
-        this.motionZ *= (double)0.98F;
+        this.motionX *= 0.98F;
+        this.motionY *= 0.98F;
+        this.motionZ *= 0.98F;
 
         if (this.particleMaxAge-- <= 0)
         {
@@ -51,8 +51,8 @@ public class EntityRainFX extends EntityFX
                 this.setDead();
             }
 
-            this.motionX *= (double)0.7F;
-            this.motionZ *= (double)0.7F;
+            this.motionX *= 0.7F;
+            this.motionZ *= 0.7F;
         }
 
         BlockPos blockpos = new BlockPos(this);
@@ -67,7 +67,7 @@ public class EntityRainFX extends EntityFX
 
             if (iblockstate.getBlock() instanceof BlockLiquid)
             {
-                d0 = (double)(1.0F - BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL)));
+                d0 = 1.0F - BlockLiquid.getLiquidHeightPercent(iblockstate.getValue(BlockLiquid.LEVEL));
             }
             else
             {
